requirements.txt
Django==1.10.6
djangorestframework==3.6.0
coreapi==2.3.0
django-rest-swagger==2.1.2
openapi-codec==1.3.1
django-cors-headers
simplejson==3.9.0
coreapi==2.3.0
openapi-codec==1.3.1
simplejson==3.9.0
django-celery-inspect
Markdown==2.6.5
PyYAML==3.11
wheel==0.24.0
cm_api==16.0.0
django-templated-email
django-anymail
unique
arrow
sklearn
scipy
matplotlib
cssselect
tinycss
lxml
django-crontab
pygal
pytz
pyhs2